Move featureflags from applications -> core
This was the only endpoint in applications that core actually uses.
