+++ api/applications/applications.proto
@@ -95,15 +95,6 @@ service Applications {
             body: "*"
         };
     }
-
-    /**
-    * Config returns configuration information about the server
-    */
-    rpc GetFeatureFlags(GetFeatureFlagsRequest) returns (GetFeatureFlagsResponse) {
-        option (google.api.http) = {
-            get : "/v1/featureflags"
-        };
-    }
 }
 
 message AuthenticateRequest {
@@ -174,10 +165,4 @@ message ValidateProviderTokenRequest {
 
 message ValidateProviderTokenResponse {
     bool valid = 1;
-}
-
-message GetFeatureFlagsRequest {}
-
-message GetFeatureFlagsResponse {
-    map<string, string> flags = 1;
 }
