+++ api/applications/applications.swagger.json
@@ -244,29 +244,6 @@
           "Applications"
         ]
       }
-    },
-    "/v1/featureflags": {
-      "get": {
-        "summary": "Config returns configuration information about the server",
-        "operationId": "Applications_GetFeatureFlags",
-        "responses": {
-          "200": {
-            "description": "A successful response.",
-            "schema": {
-              "$ref": "#/definitions/v1GetFeatureFlagsResponse"
-            }
-          },
-          "default": {
-            "description": "An unexpected error response.",
-            "schema": {
-              "$ref": "#/definitions/rpcStatus"
-            }
-          }
-        },
-        "tags": [
-          "Applications"
-        ]
-      }
     }
   },
   "definitions": {
@@ -321,17 +298,6 @@
       "properties": {
         "token": {
           "type": "string"
-        }
-      }
-    },
-    "v1GetFeatureFlagsResponse": {
-      "type": "object",
-      "properties": {
-        "flags": {
-          "type": "object",
-          "additionalProperties": {
-            "type": "string"
-          }
         }
       }
     },
