+++ api/core/core.proto
@@ -163,6 +163,15 @@ service Core {
             body: "*"
         };
     }
+
+    /**
+    * GetFeatureFlags returns configuration information about the server
+    */
+    rpc GetFeatureFlags(GetFeatureFlagsRequest) returns (GetFeatureFlagsResponse) {
+        option (google.api.http) = {
+            get : "/v1/featureflags"
+        };
+    }
 }
 
 message Pagination {
@@ -302,4 +311,10 @@ message SyncAutomationRequest {
 }
 
 message SyncAutomationResponse {
+}
+
+message GetFeatureFlagsRequest {}
+
+message GetFeatureFlagsResponse {
+    map<string, string> flags = 1;
 }
